perf: force update state after snap request (#24560)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Adds logic to force update the MetaMask state after requests to Snaps
that create interfaces. This reduces the latency of the first viewing of
the home page as the state is synced instantly and not debounced.